+++ src/main/java/com/chatop/repository/DBUserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DBUserRepository extends JpaRepository<DBUser, Long> {
+public interface DBUserRepository extends JpaRepository<DBUser, Integer> {
     public Optional<DBUser> findByEmail(String email);
 }
